requirements.txt
setuptools>=70.0  # provides distutils, removed in Python 3.12
openai>=1.37.0
selenium>=4.9.1
undetected-chromedriver>=3.5.5
webdriver-manager==4.0.2
httpx~=0.27.2
Levenshtein==0.25.1
PyYAML>=6.0
loguru>=0.7.2
redis>=5.0
